[uievents] Simplify dblclick test to make it fail fast
Previously, if the actions_promise was rejected, which happens when
running manually, the test would simply time out. More importantly, it
makes it harder to know why Safari is failing:
https://wpt.fyi/results/uievents/click/mouse-dblclick-event.html?run_id=536660003&run_id=515490001&run_id=532760001

This also renames the test to match sibling tests and make room for a
dblclick_event_touch.html to be added.