bluetooth: implement writeValueWith[out]Response()
This implements the Web Bluetooth BluetoothRemoteGATTCharacteristic
writeValueWithResponse() and writeValueWithoutResponse() methods.

https://webbluetoothcg.github.io/web-bluetooth/#writecharacteristicvalue

The writeValue() method is now deprecated and should continue to work
exactly as it did before.

New web_tests are just a copy of existing tests for writeValue(). The
tests for writeValueWithResponse() are exactly the same. The tests for
writeValueWithoutResponse() have calls to setNextWriteResponse()
removed.

Bug: 672648
Change-Id: I0d8032fded0028eaa78a55afe3336f253c85c590