Added in special case for timeline inclusive end point.
A special case is needed for timeline current time and phase calculation
when endScrollOffset is greater than or equal to the max offset. This
special case was removed when adding timeline phase, and is now being
added back in.

The special case makes it so that when no endScrollOffset is provided
(which is the default behavior) the max scroll offset is inclusive.
This means that it is not necessary to specify "fill: forwards" on an
effect to make sure it remains at the end of the effect when the scroll
has reached its maximum.

https://github.com/w3c/csswg-drafts/issues/4325#issuecomment-606723725

Bug: 1046833
Change-Id: I6184f9c6059e1a091b54d86f37e4ba4b63c6bd13