[FlexNG] Clear measure result if layout dirty, and layout result is set.
There was a subtle bug with the measure layout result.

We could get into a state previously where the "measure" layout result,
was invalid (pointing to a previous layout). We'd then hit the cache
using the "measure" slot, causing a UAF.

This patch fixes this issue in two different ways:
1) Clears the measure slot when we are setting a layout result, and we
   needed layout.
2) Within ComputeMinMaxSizes use the "measure" cache-slot.

An additional subtle issue was uncovered while investigating this, where
it was possible to hit an incorrect layout result.

Now the "measure" cache-slot should only be used, if it is guaranteed
that a subsequent layout will occur for that node.

Bug: 1081086
Change-Id: I08267eadedc7599075e4b6867d341223999ad9f6