Update Storage Access API Logic & E2E Tests
This change updates the logic used in the Blink document for
document.hasStorageAccess and document.requestStorageAccess. Gates have
been added to ensure the proper restrictions such as having a non-opaque
origin, sandbox token (if applicable), as well as ensuring that
hasStorageAccess reflects the actual state of storage access on the page.

Additionally defense in depth logic checks have been added in
StorageAccessGrantPermissionContext to shortcut if we get a request for
a scenario no grant would be needed.

Some additional cleanup is included with this change to remove the
temporary usage of CONTENT_SETTING_SESSION in
StorageAccessGrantPermissionContext and to ensure all content settings
updates will propagate to the network service. A follow up change will
ensure this does not result in duplicate updates when we perform our
initial update.

Browser tests have been added to validate end to end functionality.

Bug: 989663
Change-Id: I09fdc115a30a71e341961b385a2e2e05fd161536